feat: 템플스테이 상세페이지 SSR 전환
